+++ frontend/src/lib/api/ckbtc-ledger.api.ts
@@ -6,40 +6,48 @@ import {
   type IcrcTransferParams,
 } from "$lib/api/icrc-ledger.api";
 import { HOST } from "$lib/constants/environment.constants";
-import type { Account } from "$lib/types/account";
+import type { Account, AccountType } from "$lib/types/account";
 import type { IcrcTokenMetadata } from "$lib/types/icrc";
 import { logWithTimestamp } from "$lib/utils/dev.utils";
 import type { HttpAgent, Identity } from "@dfinity/agent";
-import { IcrcLedgerCanister } from "@dfinity/ledger";
+import {
+  IcrcLedgerCanister,
+  type IcrcAccount,
+  type IcrcBlockIndex,
+} from "@dfinity/ledger";
 import type { Principal } from "@dfinity/principal";
 
-export const getCkBTCAccounts = async ({
+export const getCkBTCAccount = async ({
   identity,
   certified,
   canisterId,
+  ...rest
 }: {
   identity: Identity;
   certified: boolean;
   canisterId: Principal;
-}): Promise<Account[]> => {
-  // TODO: Support subaccounts
-  logWithTimestamp("Getting ckBTC accounts: call...");
+  type: AccountType;
+} & IcrcAccount): Promise<Account> => {
+  logWithTimestamp("Getting ckBTC account: call...");
 
   const {
     canister: { metadata, balance },
   } = await ckBTCLedgerCanister({ identity, canisterId });
 
-  const mainAccount = await getIcrcAccount({
-    owner: identity.getPrincipal(),
-    type: "main",
+  const callParams = {
     certified,
     getBalance: balance,
     getMetadata: metadata,
+  };
+
+  const account = await getIcrcAccount({
+    ...rest,
+    ...callParams,
   });
 
-  logWithTimestamp("Getting ckBTC accounts: done");
+  logWithTimestamp("Getting ckBTC account: done");
 
-  return [mainAccount];
+  return account;
 };
 
 export const getCkBTCToken = async ({
@@ -74,19 +82,21 @@ export const ckBTCTransfer = async ({
 }: {
   identity: Identity;
   canisterId: Principal;
-} & Omit<IcrcTransferParams, "transfer">): Promise<void> => {
+} & Omit<IcrcTransferParams, "transfer">): Promise<IcrcBlockIndex> => {
   logWithTimestamp("Getting ckBTC transfer: call...");
 
   const {
     canister: { transfer: transferApi },
   } = await ckBTCLedgerCanister({ identity, canisterId });
 
-  await transferIcrcApi({
+  const blockIndex = await transferIcrcApi({
     ...rest,
     transfer: transferApi,
   });
 
   logWithTimestamp("Getting ckBTC transfer: done");
+
+  return blockIndex;
 };
 
 const ckBTCLedgerCanister = async ({