+++ frontend/src/lib/api/sns-ledger.api.ts
@@ -8,6 +8,7 @@ import type { Account } from "$lib/types/account";
 import type { IcrcTokenMetadata } from "$lib/types/icrc";
 import { logWithTimestamp } from "$lib/utils/dev.utils";
 import type { Identity } from "@dfinity/agent";
+import type { IcrcBlockIndex } from "@dfinity/ledger";
 import type { Principal } from "@dfinity/principal";
 import { wrapper } from "./sns-wrapper.api";
 
@@ -98,7 +99,7 @@ export const snsTransfer = async ({
 }: {
   identity: Identity;
   rootCanisterId: Principal;
-} & Omit<IcrcTransferParams, "transfer">): Promise<void> => {
+} & Omit<IcrcTransferParams, "transfer">): Promise<IcrcBlockIndex> => {
   logWithTimestamp("Getting Sns transfer: call...");
 
   const { transfer: transferApi } = await wrapper({
@@ -107,10 +108,12 @@ export const snsTransfer = async ({
     certified: true,
   });
 
-  await transferIcrcApi({
+  const blockIndex = await transferIcrcApi({
     ...rest,
     transfer: transferApi,
   });
 
   logWithTimestamp("Getting Sns transfer: done");
+
+  return blockIndex;
 };