+++ frontend/src/lib/components/warnings/Warnings.svelte
@@ -1,10 +1,12 @@
 <script lang="ts">
-  import TransactionRateWarning from "$lib/components/metrics/TransactionRateWarning.svelte";
+  import TransactionRateWarning from "$lib/components/warnings/TransactionRateWarning.svelte";
   import { ENABLE_METRICS } from "$lib/constants/mockable.constants";
   import Metrics from "$lib/components/metrics/Metrics.svelte";
   import { Toasts } from "@dfinity/gix-components";
+  import ConvertCkBTCToBtcWarning from "$lib/components/warnings/ConvertCkBTCToBtcWarning.svelte";
 
   export let bringToastsForward = false;
+  export let ckBTCWarnings = false;
 </script>
 
 {#if ENABLE_METRICS}
@@ -16,6 +18,10 @@
 </div>
 
 <TransactionRateWarning />
+
+{#if ckBTCWarnings}
+  <ConvertCkBTCToBtcWarning />
+{/if}
 
 <style lang="scss">
   .forward {