+++ frontend/src/lib/constants/stores.constants.ts
@@ -1,5 +1,6 @@
-export enum storeLocalStorageKey {
+export enum StoreLocalStorageKey {
   ProposalFilters = "nnsProposalFilters",
   Theme = "nnsTheme",
   FeatureFlags = "nnsOverrideFeatureFlags",
+  BitcoinConvertBlockIndexes = "nnsBitcoinConvertBlockIndexes",
 }