+++ frontend/src/lib/modals/accounts/CkBTCTransactionModal.svelte
@@ -68,7 +68,7 @@
       initiator: "accounts",
     });
 
-    const { success } = await ckBTCTransferTokens({
+    const { blockIndex } = await ckBTCTransferTokens({
       source: sourceAccount,
       destinationAddress,
       amount,
@@ -79,7 +79,7 @@
 
     stopBusy("accounts");
 
-    if (success) {
+    if (nonNullish(blockIndex)) {
       toastsSuccess({ labelKey: "accounts.transaction_success" });
       dispatcher("nnsTransfer");
     }