+++ frontend/src/lib/services/ckbtc-convert.services.ts
@@ -1,7 +1,8 @@
+import { getWithdrawalAccount } from "$lib/services/ckbtc-minter.services";
+import { bitcoinConvertBlockIndexes } from "$lib/stores/bitcoin.store";
 import type { CkBTCAdditionalCanisters } from "$lib/types/ckbtc-canisters";
 import { ConvertBtcStep } from "$lib/types/ckbtc-convert";
 import type { UniverseCanisterId } from "$lib/types/universe";
-import type { WithdrawalAccount } from "@dfinity/ckbtc";
 import {
   MinterAlreadyProcessingError,
   MinterAmountTooLowError,
@@ -11,10 +12,9 @@ import {
   MinterTemporaryUnavailableError,
 } from "@dfinity/ckbtc";
 import { encodeIcrcAccount } from "@dfinity/ledger";
-import { assertNonNullish, fromNullable } from "@dfinity/utils";
-import { getWithdrawalAccount, retrieveBtc } from "../api/ckbtc-minter.api";
+import { fromNullable, isNullish } from "@dfinity/utils";
+import { retrieveBtc } from "../api/ckbtc-minter.api";
 import { toastsError } from "../stores/toasts.store";
-import { toToastError } from "../utils/error.utils";
 import { numberToE8s } from "../utils/token.utils";
 import { getAuthenticatedIdentity } from "./auth.services";
 import { ckBTCTransferTokens } from "./ckbtc-accounts.services";
@@ -48,26 +48,11 @@ export const convertCkBTCToBtc = async ({
 
   const identity = await getAuthenticatedIdentity();
 
-  let account: WithdrawalAccount | undefined;
+  const account = await getWithdrawalAccount({ minterCanisterId });
 
-  try {
-    account = await getWithdrawalAccount({
-      identity,
-      canisterId: minterCanisterId,
-    });
-  } catch (err: unknown) {
-    toastsError(
-      toToastError({
-        err,
-        fallbackErrorLabelKey: "error__ckbtc.withdrawal_account",
-      })
-    );
-
+  if (isNullish(account)) {
     return { success: false };
   }
-
-  // Account cannot be null here. We add this guard to comply with type safety.
-  assertNonNullish(account);
 
   // For simplicity and compatibility reason with the transferTokens interface we just encode the account here instead of extending the interface to support either account identifier or {owner; subaccount;} object.
   const ledgerAddress = encodeIcrcAccount({
@@ -78,7 +63,7 @@ export const convertCkBTCToBtc = async ({
   updateProgress(ConvertBtcStep.LOCKING_CKBTC);
 
   // We reload the transactions only at the end of the process for performance reason.
-  const { success: transferSuccess } = await ckBTCTransferTokens({
+  const { blockIndex } = await ckBTCTransferTokens({
     source,
     amount,
     destinationAddress: ledgerAddress,
@@ -87,11 +72,12 @@ export const convertCkBTCToBtc = async ({
     indexCanisterId,
   });
 
-  if (!transferSuccess) {
+  if (isNullish(blockIndex)) {
     return { success: false };
   }
 
-  // TODO(GIX-1324): how do we handle failure between these steps UX wise?
+  // Flag the transaction that was executed
+  bitcoinConvertBlockIndexes.addBlockIndex(blockIndex);
 
   updateProgress(ConvertBtcStep.SEND_BTC);
 
@@ -107,6 +93,9 @@ export const convertCkBTCToBtc = async ({
 
     return { success: false };
   } finally {
+    // Regardless if success or error, the UI is still active therefore we can remove the flag
+    bitcoinConvertBlockIndexes.removeBlockIndex(blockIndex);
+
     updateProgress(ConvertBtcStep.RELOAD);
 
     await loadCkBTCAccountTransactions({