+++ frontend/src/tests/lib/services/ckbtc-accounts.services.spec.ts
@@ -38,8 +38,8 @@ describe("ckbtc-accounts-services", () => {
 
     it("should call api.getSnsAccounts and load neurons in store", async () => {
       const spyQuery = jest
-        .spyOn(ledgerApi, "getCkBTCAccounts")
-        .mockImplementation(() => Promise.resolve([mockCkBTCMainAccount]));
+        .spyOn(ledgerApi, "getCkBTCAccount")
+        .mockImplementation(() => Promise.resolve(mockCkBTCMainAccount));
 
       await loadCkBTCAccounts({ universeId: CKBTC_UNIVERSE_CANISTER_ID });
 
@@ -57,7 +57,7 @@ describe("ckbtc-accounts-services", () => {
 
     it("should call error callback", async () => {
       const spyQuery = jest
-        .spyOn(ledgerApi, "getCkBTCAccounts")
+        .spyOn(ledgerApi, "getCkBTCAccount")
         .mockRejectedValue(new Error());
 
       const spy = jest.fn();
@@ -90,7 +90,7 @@ describe("ckbtc-accounts-services", () => {
       });
 
       jest
-        .spyOn(ledgerApi, "getCkBTCAccounts")
+        .spyOn(ledgerApi, "getCkBTCAccount")
         .mockImplementation(() => Promise.reject(undefined));
 
       await loadCkBTCAccounts({ universeId: CKBTC_UNIVERSE_CANISTER_ID });
@@ -113,8 +113,8 @@ describe("ckbtc-accounts-services", () => {
 
     it("should call ckBTC accounts and token and load them in store", async () => {
       const spyAccountsQuery = jest
-        .spyOn(ledgerApi, "getCkBTCAccounts")
-        .mockImplementation(() => Promise.resolve([mockCkBTCMainAccount]));
+        .spyOn(ledgerApi, "getCkBTCAccount")
+        .mockImplementation(() => Promise.resolve(mockCkBTCMainAccount));
 
       const spyTokenQuery = jest
         .spyOn(ledgerApi, "getCkBTCToken")
@@ -147,8 +147,8 @@ describe("ckbtc-accounts-services", () => {
 
   describe("ckBTCTransferTokens", () => {
     const spyAccounts = jest
-      .spyOn(ledgerApi, "getCkBTCAccounts")
-      .mockImplementation(() => Promise.resolve([mockCkBTCMainAccount]));
+      .spyOn(ledgerApi, "getCkBTCAccount")
+      .mockImplementation(() => Promise.resolve(mockCkBTCMainAccount));
 
     beforeEach(() => {
       jest.clearAllMocks();
@@ -164,9 +164,9 @@ describe("ckbtc-accounts-services", () => {
 
       const spyTransfer = jest
         .spyOn(ledgerApi, "ckBTCTransfer")
-        .mockResolvedValue(undefined);
+        .mockResolvedValue(456n);
 
-      const { success } = await services.ckBTCTransferTokens({
+      const { blockIndex } = await services.ckBTCTransferTokens({
         source: mockCkBTCMainAccount,
         destinationAddress: "aaaaa-aa",
         amount: 1,
@@ -175,7 +175,7 @@ describe("ckbtc-accounts-services", () => {
         indexCanisterId: CKBTC_INDEX_CANISTER_ID,
       });
 
-      expect(success).toBe(true);
+      expect(blockIndex).toEqual(456n);
       expect(spyTransfer).toBeCalled();
       expect(spyAccounts).toBeCalled();
     });
@@ -185,9 +185,9 @@ describe("ckbtc-accounts-services", () => {
 
       const spyTransfer = jest
         .spyOn(ledgerApi, "ckBTCTransfer")
-        .mockResolvedValue(undefined);
+        .mockResolvedValue(456n);
 
-      const { success } = await services.ckBTCTransferTokens({
+      const { blockIndex } = await services.ckBTCTransferTokens({
         source: mockSnsMainAccount,
         destinationAddress: "aaaaa-aa",
         amount: 1,
@@ -196,7 +196,7 @@ describe("ckbtc-accounts-services", () => {
         indexCanisterId: CKBTC_INDEX_CANISTER_ID,
       });
 
-      expect(success).toBe(true);
+      expect(blockIndex).toEqual(456n);
       expect(spyTransfer).toBeCalled();
       expect(spyAccounts).toBeCalled();
       expect(loadCkBTCAccountTransactions).toBeCalled();
@@ -212,7 +212,7 @@ describe("ckbtc-accounts-services", () => {
 
       const spyOnToastsError = jest.spyOn(toastsStore, "toastsError");
 
-      const { success } = await services.ckBTCTransferTokens({
+      const { blockIndex } = await services.ckBTCTransferTokens({
         source: mockSnsMainAccount,
         destinationAddress: "aaaaa-aa",
         amount: 1,
@@ -221,7 +221,7 @@ describe("ckbtc-accounts-services", () => {
         indexCanisterId: CKBTC_INDEX_CANISTER_ID,
       });
 
-      expect(success).toBe(false);
+      expect(blockIndex).toBeUndefined();
       expect(spyTransfer).toBeCalled();
       expect(spyAccounts).not.toBeCalled();
       expect(spyOnToastsError).toBeCalled();
@@ -236,7 +236,7 @@ describe("ckbtc-accounts-services", () => {
 
       const spyOnToastsError = jest.spyOn(toastsStore, "toastsError");
 
-      const { success } = await services.ckBTCTransferTokens({
+      const { blockIndex } = await services.ckBTCTransferTokens({
         source: mockSnsMainAccount,
         destinationAddress: "aaaaa-aa",
         amount: 1,
@@ -245,7 +245,7 @@ describe("ckbtc-accounts-services", () => {
         indexCanisterId: CKBTC_INDEX_CANISTER_ID,
       });
 
-      expect(success).toBe(false);
+      expect(blockIndex).toBeUndefined();
       expect(spyTransfer).not.toBeCalled();
       expect(spyAccounts).not.toBeCalled();
       expect(spyOnToastsError).toBeCalled();